docs(core): move metrics to telemetry namespace
