+++ core/samples/metrics/main.cpp
@@ -7,9 +7,9 @@
 #include <cubos/core/tel/logging.hpp>
 
 /// [Import metrics]
-#include <cubos/core/metrics.hpp>
+#include <cubos/core/tel/metrics.hpp>
 
-using cubos::core::Metrics;
+using cubos::core::tel::Metrics;
 /// [Import metrics]
 
 /// [Generate random value]
